requirements.txt
# Core
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
httpx>=0.27.0
rich>=13.7.0
tenacity>=8.2.3

# Content extraction
trafilatura>=1.8.0
lxml_html_clean

# Search providers
ddgs>=6.1.0
tavily-python>=0.3.6

# LLMs
mistralai>=1.2.2
google-generativeai>=0.7.2

# Orchestration
langgraph>=0.2.28
langchain-core>=0.2.39

# MCP Protocol (SDK for future server endpoints)
mcp

# Utils
json-repair>=0.27.0